got webpack to output both chrome and firefox versions, update build script - ms
