+++ public/manifest.chrome.json
@@ -1,7 +1,7 @@
 {
   "name": "Voice DIM",
   "description": "Control DIM with your voice.",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "manifest_version": 3,
   "background": {
     "service_worker": "js/background.js"
