+++ public/manifest.firefox.json
@@ -1,7 +1,7 @@
 {
   "name": "DIM Voice",
   "description": "Control DIM with your voice.",
-  "version": "1.0.0",
+  "version": "1.0.2",
   "manifest_version": 2,
   "background": {
     "scripts": ["js/background.js"]
