+++ webpack/webpack.prod.js
@@ -1,6 +1,10 @@
 const { merge } = require('webpack-merge');
 const common = require('./webpack.common.js');
 
-module.exports = merge(common, {
-  mode: 'production',
-});
+module.exports = (env) => {
+  console.log({ common });
+  console.log({ env });
+  return merge(common, {
+    mode: 'production',
+  });
+};
